augur/core: split scenario_tax_defaults out so local_regulation stays a leaf
